feat(deconv): add extra term in denominator for finite mean
This is a naive first implementation of an addition to the deconvolution
algorithm that tries to fix the issue of deconvolving signals with
unknown and non-zero mean. The implementation is from a manuscript
describing "Finite noise mean", but whether this should be added to the
denominator or subtracted from the numerator is not yet clear.

refactor(deconv): give the added term better name